fix: PR 코멘트 대응 - FSD 위반, 경로 하드코딩, 에러 처리, 색상 토큰 수정
- TestCard를 shared/ui로 이동하여 features 간 직접 의존 해소
- features/discovery/ui/TestCard는 shared/ui/TestCard re-export로 대체
- InterestList: ROUTES.INTEREST_DETAIL 사용으로 경로 하드코딩 제거
- InterestList: #4365cc → adaptive.blue600 (TDS 색상 토큰)
- interest/$testId: isError 처리 추가 및 Result 에러 UI 렌더링
